requirements.txt
langchain
ipykernel
python-dotenv
langchain-community
